isobmff: fix issue on test
Tested:
```
$ make test
Running tests...
Test project isobmff/build
      Start  1: AVC1_unittest
 1/21 Test  #1: AVC1_unittest ....................   Passed    0.02 sec
      Start  2: AVC3_unittest
 2/21 Test  #2: AVC3_unittest ....................   Passed    0.00 sec
      Start  3: AVCC_unittest
 3/21 Test  #3: AVCC_unittest ....................   Passed    0.00 sec
      Start  4: CTTS_unittest
 4/21 Test  #4: CTTS_unittest ....................   Passed    0.00 sec
      Start  5: DREF_unittest
 5/21 Test  #5: DREF_unittest ....................   Passed    0.00 sec
      Start  6: FTYP_unittest
 6/21 Test  #6: FTYP_unittest ....................   Passed    0.00 sec
      Start  7: HDLR_unittest
 7/21 Test  #7: HDLR_unittest ....................   Passed    0.00 sec
      Start  8: HEV1_unittest
 8/21 Test  #8: HEV1_unittest ....................   Passed    0.00 sec
      Start  9: HVC1_unittest
 9/21 Test  #9: HVC1_unittest ....................   Passed    0.00 sec
      Start 10: HVCC_unittest
10/21 Test #10: HVCC_unittest ....................   Passed    0.00 sec
      Start 11: ISOBMFF_memory_test
11/21 Test #11: ISOBMFF_memory_test ..............   Passed    0.03 sec
      Start 12: MDHD_unittest
12/21 Test #12: MDHD_unittest ....................   Passed    0.00 sec
      Start 13: META_unittest
13/21 Test #13: META_unittest ....................   Passed    0.00 sec
      Start 14: MP4A_unittest
14/21 Test #14: MP4A_unittest ....................   Passed    0.00 sec
      Start 15: MVHD_unittest
15/21 Test #15: MVHD_unittest ....................   Passed    0.00 sec
      Start 16: Parser_unittest
16/21 Test #16: Parser_unittest ..................   Passed    0.00 sec
      Start 17: STSD_unittest
17/21 Test #17: STSD_unittest ....................   Passed    0.00 sec
      Start 18: STSS_unittest
18/21 Test #18: STSS_unittest ....................   Passed    0.00 sec
      Start 19: STTS_unittest
19/21 Test #19: STTS_unittest ....................   Passed    0.00 sec
      Start 20: TKHD_unittest
20/21 Test #20: TKHD_unittest ....................   Passed    0.00 sec
      Start 21: URL_unittest
21/21 Test #21: URL_unittest .....................   Passed    0.00 sec

100% tests passed, 0 tests failed out of 21

Total Test time (real) =   0.13 sec
```

+++ test/ISOBMFF_memory_test.cpp
@@ -73,7 +73,7 @@ class IsobmffMemoryTest : public ::testing::Test {
       return 0;
     }
     fclose(fp);
-    return rss * sysconf(_SC_PAGESIZE);
+    return static_cast<size_t>(rss) * static_cast<size_t>(sysconf(_SC_PAGESIZE));
 #else
     return 0;
 #endif
